security/fbfam: Add two new prctls to enable and disable the fbfam feature
To allow that a process can turn off or turn on the detection and
mitigation of a fork brute force attack when required, add two new
defines to the prctl interface.

All the arguments passed to the prctl system call are ignored for the
two new cases. It returns zero on success and -EFAULT if the current
task doesn't have statistical data.

Signed-off-by: John Wood <john.wood@gmx.com>