securtiy/brute: Detect a brute force attack
To detect a fork brute force attack it is necessary that the statistics
shared by all the fork hierarchy processes be updated in every fatal
crash and the most important data to update is the application crash
period. To do so, use the new "task_fatal_signal" LSM hook added in a
previous step.

The application crash period must be a value that is not prone to change
due to spurious data and follows the real crash period. So, to compute
it, the exponential moving average (EMA) is used.

A fork brute force attack has two variants. A slow brute force attack
that is detected if the maximum number of faults per fork hierarchy is
reached and a fast brute force attack that is detected if the
application crash period falls below a certain threshold.

Signed-off-by: John Wood <john.wood@gmx.com>